Changes: -Added a global header with App Name -Edited how cars appear on the list -Added "Edit" button for Garages -Added an alert before Delete/Remove operations

+++ Pages/Cars.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
-import { Alert, ScrollView, Text, TextInput, TouchableOpacity, View, } from 'react-native';
+import { Alert, ScrollView, StyleSheet, Text, TextInput, TouchableOpacity, View, } from 'react-native';
 import { Picker } from '@react-native-picker/picker';
 import AsyncStorage from '@react-native-community/async-storage';
 
-const Cars = ({ addGarage }) => {
+const Cars = () => {
 
   const [garageNames, setGarageNames] = useState([]);
   const [allCars, setAllCars] = useState([]);
@@ -11,6 +11,7 @@ const Cars = ({ addGarage }) => {
   const [selectedGarage, setSelectedGarage] = useState('');
 
   useEffect(() => {
+    // Get the list of garage names from local storage
     const getGarageNames = async () => {
       const garageNames = await retrieveData('garageNames');
       setGarageNames(garageNames);
@@ -19,6 +20,7 @@ const Cars = ({ addGarage }) => {
   }, []);
 
   useEffect(() => {
+    // Get the list of all cars in all garages from local storage
     const getCars = async () => {
       const cars = [];
       for (const garageName of garageNames) {
@@ -31,21 +33,20 @@ const Cars = ({ addGarage }) => {
     getCars();
   }, [garageNames]);
 
-  const addNewGarage = async () => {
-    await addGarage();
-    const updatedGarageNames = await retrieveData('garageNames');
-    setGarageNames(updatedGarageNames);
+  const updateGarageNames = async () => {
+    const garageNames = await retrieveData('garageNames');
+    setGarageNames(garageNames);
   };
 
   const addNewCar = async () => {
 
     if (!carName) {
-      Alert.alert('Error', "Car name can not be empty.");
+      Alert.alert('Add New Car', "Car name can not be empty.");
       return;
     }
 
     if (carName.includes(']')) {
-      Alert.alert('Error', 'Car name can not contain " ] " character.');
+      Alert.alert('Add New Car', 'Car name can not contain " ] " character.');
       return;
     }
 
@@ -67,39 +68,57 @@ const Cars = ({ addGarage }) => {
       allCarNames.sort();
       setAllCars(allCarNames);
 
+      setCarName('');
+
     } catch (error) {
       console.error(error);
     }
   };
 
-  const deleteCar = async (carNameWithGarageName) => {
+  const removeCar = async (carNameWithGarageName) => {
 
-    try {
+    const garageNameEndIndex = carNameWithGarageName.indexOf(']');
+    const nameofTheGarage = carNameWithGarageName.substr(1, garageNameEndIndex - 1);
+    const carToRemove = carNameWithGarageName.substr(garageNameEndIndex + 2);
 
-      const garageNameEndIndex = carNameWithGarageName.indexOf(']');
-      const nameofTheGarage = carNameWithGarageName.substr(1, garageNameEndIndex - 1);
-      const carToRemove = carNameWithGarageName.substr(garageNameEndIndex + 2);
+    Alert.alert(
+      'Remove Car',
+      'Are you sure you want to remove ' + carToRemove + ' in ' + nameofTheGarage + '?',
+      [
+        {
+          text: 'Cancel',
+          style: 'cancel',
+        },
+        {
+          text: 'OK',
+          onPress: async () => {
+            try {
 
-      // Find and remove the car from specific garage
-      const garageCars = await retrieveData(`${nameofTheGarage}_cars`);
-      const updatedGarageCars = garageCars.filter(e => e !== carNameWithGarageName);
+              // Find and remove the car from specific garage
+              const garageCars = await retrieveData(`${nameofTheGarage}_cars`);
+              const updatedGarageCars = garageCars.filter(e => e !== carNameWithGarageName);
 
-      await AsyncStorage.setItem(`${nameofTheGarage}_cars`, JSON.stringify(updatedGarageCars));
+              await AsyncStorage.setItem(`${nameofTheGarage}_cars`, JSON.stringify(updatedGarageCars));
 
-      // Update the car list
-      const allCarNames = [];
+              // Update the car list
+              const allCarNames = [];
 
-      for (const garageName of garageNames) {
-        const garageCars = await retrieveData(`${garageName}_cars`);
-        allCarNames.push(...garageCars);
-      }
+              for (const garageName of garageNames) {
+                const garageCars = await retrieveData(`${garageName}_cars`);
+                allCarNames.push(...garageCars);
+              }
 
-      allCarNames.sort();
-      setAllCars(allCarNames);
+              allCarNames.sort();
+              setAllCars(allCarNames);
 
-    } catch (error) {
-      console.error(error);
-    }
+            } catch (error) {
+              console.error(error);
+            }
+          },
+        },
+      ],
+      { cancelable: false }
+    );
   };
 
   return (
@@ -107,38 +126,34 @@ const Cars = ({ addGarage }) => {
       <ScrollView>
         {allCars.map((carName, index) => (
           <View
-            style={{
-              flexDirection: 'row',
-              alignItems: 'center',
-              padding: 16,
-              borderBottomWidth: 1,
-              borderBottomColor: '#ccc'
-            }}
+            style={styles.carListContainer}
           >
 
-          <TouchableOpacity>
-            <Text style={{color: 'black'}}>{carName}</Text>
-          </TouchableOpacity>
+          <View style={{ flex: 1, flexDirection:'row' }}>
+            <TouchableOpacity>
+              <Text style={{color: 'black', fontWeight: 'bold'}}>
+                {carName.substr(carName.indexOf(']') + 2)}
+                </Text>
+            </TouchableOpacity>
 
-          <TouchableOpacity
-            style={{ marginLeft: 'auto' }}
-            onPress={() => deleteCar(carName)}
-          >
-            <Text style={{color: 'red'}}>Delete</Text>
-          </TouchableOpacity>
+            <TouchableOpacity
+              style={{ marginLeft: 'auto' }}>
+              <Text style={{ color: 'black' }}>{'[' + carName.substr(1, carName.indexOf(']') - 1) + ']'}</Text>
+            </TouchableOpacity>
+
+            <TouchableOpacity
+              style={{ marginLeft: 'auto' }}
+              onPress={() => removeCar(carName)}>
+              <Text style={{color: 'red'}}>Remove</Text>
+            </TouchableOpacity>
+          </View>
         </View>
       ))}
       </ScrollView>
 
       <View style={{ flexDirection: 'row', alignItems: 'center', margin: 10 }}>
         <TextInput
-          style={{
-            height: 40,
-            borderColor: 'gray',
-            color: 'black',
-            borderWidth: 1,
-            flex: 1,
-            marginRight: 10}}
+          style={styles.addNewCarContainer}
           onChangeText={text => setCarName(text)}
           value={carName}
           placeholder="Car Name"
@@ -147,30 +162,23 @@ const Cars = ({ addGarage }) => {
 
         <Picker
           selectedValue={selectedGarage}
-          style={{
-            height: 40,
-            width: 200,
-            dropdownIconColor: 'black'}}
+          style={styles.picketContainer}
           onValueChange={itemValue => setSelectedGarage(itemValue)}
           dropdownIconColor= 'black'
-          prompt='Your Garages'
-          >
+          prompt='Your Garages'>
+
             {garageNames.map(name => (
               <Picker.Item label={name} value={name} color='black' />
             ))}
+
         </Picker>
       </View>
 
       <TouchableOpacity
         onPress={addNewCar}
-        style={{
-          height: 40,
-          backgroundColor: '#009A44',
-          alignItems: 'center',
-          justifyContent: 'center',
-          margin: 10 }}
+        style={styles.button}
         >
-        <Text style={{ color: 'white' }}>Add Car</Text>
+        <Text style={{ color: 'white' }}>Add New Car</Text>
       </TouchableOpacity>
     </View>
   );
@@ -188,5 +196,40 @@ const retrieveData = async key => {
     return [];
   }
 };
+
+const styles = StyleSheet.create({
+  button: {
+    padding: 10,
+    backgroundColor: '#2D640F',
+    alignItems: 'center',
+    justifyContent: 'center',
+    margin: 10
+  },
+  addNewCarContainer: {
+    height: 40,
+    borderColor: 'gray',
+    color: 'black',
+    borderWidth: 1,
+    flex: 1,
+    marginRight: 10
+  },
+  carListContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    padding: 15,
+    borderBottomWidth: 1,
+    borderBottomColor: '#ccc'
+  },
+  picketContainer: {
+    height: 40,
+    width: 200,
+    dropdownIconColor: 'black'
+  },
+  textInput: {
+    color: 'black',
+    borderWidth:0.5,
+    margin: 10
+  }
+});
 
 export default Cars;
